Fix: LT-18378 Update Flex to use the 64-bit Hunspell Build
   * Make it the SpellEngine in single class windows and linux
